fix: 色計算ユーティリティを colorUtils.ts に共通化
hexToRgb / getLuminance / darkenHex が FilterBar と Badge に重複して
定義されていたため、src/lib/colorUtils.ts に切り出して共通化した。
