+++ src/components/FilterBar/FilterBar.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { darkenHex, getLuminance } from "@/lib/colorUtils";
+
 export type LabelOption = { name: string; color: string };
 
 export type FilterState = {
@@ -13,38 +15,6 @@ type FilterBarProps = {
   availableLabels: LabelOption[];
   onChange: (filter: FilterState) => void;
 };
-
-function hexToRgb(hex: string): { r: number; g: number; b: number } | null {
-  const cleaned = hex.replace(/^#/, "");
-  const full =
-    cleaned.length === 3
-      ? cleaned
-          .split("")
-          .map((c) => c + c)
-          .join("")
-      : cleaned;
-  const n = Number.parseInt(full, 16);
-  if (Number.isNaN(n)) return null;
-  return { r: (n >> 16) & 255, g: (n >> 8) & 255, b: n & 255 };
-}
-
-function getLuminance(hex: string): number {
-  const rgb = hexToRgb(hex);
-  if (!rgb) return 1;
-  return (0.299 * rgb.r + 0.587 * rgb.g + 0.114 * rgb.b) / 255;
-}
-
-function darkenHex(hex: string, factor: number): string {
-  const rgb = hexToRgb(hex);
-  if (!rgb) return `#${hex}`;
-  return `#${[rgb.r, rgb.g, rgb.b]
-    .map((v) =>
-      Math.round(v * factor)
-        .toString(16)
-        .padStart(2, "0"),
-    )
-    .join("")}`;
-}
 
 function getLabelStyle(color: string, active: boolean) {
   if (!active) {
