+++ src/components/ui/Badge.tsx
@@ -1,3 +1,5 @@
+import { darkenHex, getLuminance } from "@/lib/colorUtils";
+
 type BadgeVariant = "open" | "closed" | "merged" | "draft" | "custom";
 
 type BadgeProps = {
@@ -14,42 +16,6 @@ const variantClasses: Record<BadgeVariant, string> = {
   custom: "",
 };
 
-function hexToRgb(hex: string): { r: number; g: number; b: number } | null {
-  const cleaned = hex.replace(/^#/, "");
-  const full =
-    cleaned.length === 3
-      ? cleaned
-          .split("")
-          .map((c) => c + c)
-          .join("")
-      : cleaned;
-  const n = Number.parseInt(full, 16);
-  if (Number.isNaN(n)) return null;
-  return { r: (n >> 16) & 255, g: (n >> 8) & 255, b: n & 255 };
-}
-
-function luminance(hex: string): number {
-  const rgb = hexToRgb(hex);
-  if (!rgb) return 1;
-  return (0.299 * rgb.r + 0.587 * rgb.g + 0.114 * rgb.b) / 255;
-}
-
-// 明るい色を暗くしてテキストに使う
-function darkenHex(hex: string, factor: number): string {
-  const rgb = hexToRgb(hex);
-  if (!rgb) return `#${hex}`;
-  const r = Math.round(rgb.r * factor)
-    .toString(16)
-    .padStart(2, "0");
-  const g = Math.round(rgb.g * factor)
-    .toString(16)
-    .padStart(2, "0");
-  const b = Math.round(rgb.b * factor)
-    .toString(16)
-    .padStart(2, "0");
-  return `#${r}${g}${b}`;
-}
-
 export function Badge({ variant = "custom", color, children }: BadgeProps) {
   const base =
     "inline-flex items-center px-2 py-0.5 rounded text-xs font-medium";
@@ -60,7 +26,7 @@ export function Badge({ variant = "custom", color, children }: BadgeProps) {
     );
   }
 
-  const lum = luminance(color);
+  const lum = getLuminance(color);
   // 暗い色はそのまま、明るい色は 35% まで暗くしてテキストを読みやすくする
   const textColor = lum < 0.5 ? `#${color}` : darkenHex(color, 0.35);
   const style = {
